requirements.txt
pandas==2.2.1
numpy==1.26.4
nltk==3.8.1
PyPDF2==3.0.1
easyocr==1.7.1
Pillow==10.3.0
SpeechRecognition==3.10.1
gTTS==2.5.1
Flask==2.3.3
tensorflow==2.16.1
scikit-learn==1.4.2
transformers==4.40.1
torch==2.4.0
faiss-cpu==1.9.0
googletrans==4.0.0rc1
langdetect==1.0.9
reportlab==4.0.9
werkzeug==3.0.2
flask-wtf
SQLAlchemy
flask-sqlalchemy
pdfplumber
pdf2image
pytesseract
pillow
bleach
pymupdf
flask-login
